inference-api/requirements.txt
databases
fastapi
pyhumps
uvicorn
pydantic
python-dateutil
requests
sqlalchemy
sqlalchemy-utils
